+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning][].
 
 ### Added
 
+-   Pushed `get_extent` functionality upstream to `spatialdata` (#162)
+-   Multiscale image handling: user can specify a scale, else the best scale is selected automatically given the figure size and dpi (#164)
+-   Large images are automatically rasterized to speed up performance (#164)
+
 ### Fixed
 
 ## [0.0.6] - 2023-11-06